pyproject.toml
[tool.poetry]
name = "hackaithon-data"
version = "0.1.0"
description = ""
authors = ["Erick Bazán <erick.bazan@hey.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.57.0"
langchain-experimental = "^0.3.3"
langchain-community = "^0.3.9"
langchainhub = "^0.1.21"
langgraph = "^0.2.56"
langchain-openai = "^0.2.11"
llama-index-llms-openai = "^0.3.2"
llama-index-embeddings-openai = "^0.3.1"
slack-bolt = "^1.21.3"
python-dotenv = "^1.0.1"
flask = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

